requirements.txt
aiogram==3.0.0b7
sqlalchemy==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.0 